added printAll function and protected keywords
function printAll is new.

I found myself often typing pre tags and then getting the human_readable measurement. So I thought it would be nice to just say $timer->printAll('pre'); and have it all done at once. I tried to stay in your style of coding and commenting.

Also I noticed \the var keyword at the beginning of the class. Probably from PHP 4. I think it is best changed to protected. Also changed the private keywords to protected on the functions so the class can be extended. Tested and works for me.